requirements.txt
asgiref==3.7.2
azure-core==1.26.3
azure-storage-blob==12.14.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.2
Django==5.0.1
django-filter==23.5
django-storages==1.13.2
djangorestframework==3.14.0
docx==0.2.4
docx2txt==0.8
gunicorn==21.2.0
idna==3.6
install==1.3.5
isodate==0.6.1
lxml==5.1.0
Markdown==3.5.2
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
pygame==2.5.2
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tk==0.1.0
typing_extensions==4.9.0
urllib3==2.2.0
